fix(fixtures): variacion_matricula es DOUBLE PRECISION, no INTEGER
Hallazgo real de Andres al cargar el fixture: \copy fallaba porque
fact_escuela_ciclo_fixture.csv trae variacion_matricula con decimales
("0.0"), pero cargar_fixture_gold.sql la declaraba como INTEGER. Error
nuestro al escribir el loader, no del CSV ni del generador -- confirmado
contra el modelo real (dbt/models/gold/fact_escuela_ciclo.sql:54-55):
variacion_matricula se castea explicitamente a "double precision" ahi
mismo (matricula_total - matricula_ciclo_anterior as double precision),
nunca fue INTEGER en Gold real. matricula_total/matricula_ciclo_anterior
si son INTEGER (vienen de alumnos_total sin cast), esos quedan igual.

+++ tests/fixtures/gold/cargar_fixture_gold.sql
@@ -33,7 +33,7 @@ CREATE TABLE IF NOT EXISTS gold.fact_escuela_ciclo (
     id_ciclo                   TEXT,
     matricula_total            INTEGER,
     matricula_ciclo_anterior   INTEGER,
-    variacion_matricula        INTEGER,
+    variacion_matricula        DOUBLE PRECISION,
     indice_completitud_drivers DOUBLE PRECISION,
     d1 DOUBLE PRECISION,
     d2 DOUBLE PRECISION,
